+++ packages/eyes-selenium/src/SpecDriver.selenium4.js
@@ -153,6 +153,12 @@ async function isNative(driver) {
     ? ['android', 'ios'].includes(platformName.toLowerCase()) && !browserName
     : false
 }
+async function getDeviceName(driver) {
+  const capabilities = await driver.getCapabilities()
+  return capabilities.has('desired')
+    ? capabilities.get('desired').deviceName
+    : capabilities.get('deviceName')
+}
 async function getPlatformName(driver) {
   const capabilities = await driver.getCapabilities()
   return capabilities.get('platformName') || capabilities.get('platform')
@@ -260,6 +266,7 @@ exports.setWindowRect = setWindowRect
 exports.getOrientation = getOrientation
 exports.isMobile = isMobile
 exports.isNative = isNative
+exports.getDeviceName = getDeviceName
 exports.getPlatformName = getPlatformName
 exports.getPlatformVersion = getPlatformVersion
 exports.getBrowserName = getBrowserName